chore: rename tests/main.py to tests/e2emain.py
